fixed some wanrings.
git-svn-id: https://crosswire.org/svn/sword/trunk@3452 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ bindings/objc/src/SwordVerseKey.mm
@@ -130,8 +130,8 @@
     ((sword::VerseKey *)sk)->setChapter(val);
 }
 
-- (void)setVerse:(int)val {
-    ((sword::VerseKey *)sk)->setVerse(val);
+- (void)setVerse:(long)val {
+    ((sword::VerseKey *)sk)->setVerse((int)val);
 }
 
 - (NSString *)bookName {
